made the whole thing work with a proper demo. didn't realize it was broken because I am bad at github and intermittent commits.
